requirements.txt
mysql-connector==2.1.6
pyserial
